projects/roco-auto-farmer/requirements.txt
# 核心依赖
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
pyyaml>=6.0

# ADB 控制
adbutils>=2.0.0

# 图像处理
Pillow>=10.0.0

# 日志
loguru>=0.7.0

# GUI (可选)
customtkinter>=5.2.0

# 开发工具
pytest>=7.0.0
